address review comments
make sure the const eval short-circuits
removed strange re-export
