+++ compiler/plc_diagnostics/Cargo.toml
@@ -16,3 +16,6 @@ anyhow.workspace = true
 lazy_static.workspace = true
 log.workspace = true
 rustc-hash.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
