+++ src/resolver/const_evaluator.rs
@@ -495,6 +495,22 @@ fn evaluate_with_target_hint(
         }
         AstStatement::BinaryExpression(BinaryExpression { left, right, operator }) => {
             let eval_left = evaluate(left, scope, index, lhs)?;
+
+            // Short-circuit on a known boolean LHS — applies equally to AND/AND_THEN
+            // and OR/OR_ELSE. Side-stepping the RHS in const-eval prevents spurious
+            // errors like a div-by-zero on the unreachable side of `FALSE AND_THEN ...`.
+            if let Some(left_node) = eval_left.as_ref() {
+                if let AstStatement::Literal(AstLiteral::Bool(b)) = left_node.get_stmt() {
+                    match (operator, *b) {
+                        (Operator::And | Operator::AndThen, false)
+                        | (Operator::Or | Operator::OrElse, true) => {
+                            return Ok(Some(left_node.clone()));
+                        }
+                        _ => {}
+                    }
+                }
+            }
+
             let eval_right = evaluate(right, scope, index, lhs)?;
 
             if let Some((left, right)) = eval_left.zip(eval_right).as_ref() {
